splitted pipeline in files: create/enrich/viz eopatches, eolearn workflow
